+++ modules/marker/source/source.go
@@ -1,0 +1,9 @@
+package source
+
+import (
+	types "github.com/cosmos/cosmos-sdk/codec/types"
+)
+
+type Source interface {
+	GetAllMarkers(height int64) ([]*types.Any, error)
+}